requirements.txt
Django==2.1.7
djangorestframework>=3.9,<3.10
psycopg2>=2.8.4,<2.9.0
django-cors-headers
django-model-utils>=4.0.0,<4.1.0
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==19.9.0
pytz==2018.9
whitenoise==4.1.2
django-rest-swagger

